requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-dotenv>=1.0.0
faster-whisper>=1.0.0
groq>=0.4.0
edge-tts>=6.1.0
